Modernize CXXBasePaths' constructor. NFC.
As far as I can tell, the implicit initialization of DetectVirtual
to `true` in SemaDeclAttr.cpp is just an artifact of earlier
changes (and a worst-practice use of default argument values).

Actually, many of these call-sites seem to be setting values (via the
now-removed default constructor) which are irrelevant to their actual
operation. I will clean these up in a separate commit.